+++ deepqmc/solver/solver_orbital_distributed.py
@@ -135,7 +135,7 @@ class DistSolverOrbital(SolverOrbital):
             cumulative_loss = 0
             for data in self.dataloader:
 
-                lpos = Variable(data).float()
+                lpos = Variable(data)
                 lpos.requires_grad = True
 
                 loss = self.loss(lpos)
